--- ReactWindows/ReactNative.Shared/ReactNativeAppHostExtensions.cs
@@ -1,94 +1,0 @@
-using Newtonsoft.Json.Linq;
-using System;
-
-namespace ReactNative
-{
-    /// <summary>
-    /// An application delegate for managing the lifecycle events in React Native.
-    /// </summary>
-    public static class ReactNativeAppHostExtensions
-    {
-        /// <summary>
-        /// Resumes the React instance manager.
-        /// </summary>
-        /// <param name="host">The React Native host.</param>
-        /// <param name="onBackPressed">
-        /// The action to take when back is pressed.
-        /// </param>
-        public static void OnResume(this ReactNativeAppHost host, Action onBackPressed)
-        {
-            host.ReactInstanceManager.OnResume(onBackPressed);
-        }
-
-        /// <summary>
-        /// Suspends the React instance manager.
-        /// </summary>
-        /// <param name="host">The React Native host.</param>
-        public static void OnSuspend(this ReactNativeAppHost host)
-        {
-            if (host.HasInstance)
-            {
-                host.ReactInstanceManager.OnSuspend();
-            }
-        }
-
-        /// <summary>
-        /// Informs the React instance manager that the application entered
-        /// background mode.
-        /// </summary>
-        /// <param name="host">The React Native host.</param>
-        public static void OnEnteredBackground(this ReactNativeAppHost host)
-        {
-            if (host.HasInstance)
-            {
-                host.ReactInstanceManager.OnEnteredBackground();
-            }
-        }
-
-        /// <summary>
-        /// Informs the React instance manager that the application is leaving
-        /// background mode.
-        /// </summary>
-        /// <param name="host">The React Native host.</param>
-        public static void OnLeavingBackground(this ReactNativeAppHost host)
-        {
-            if (host.HasInstance)
-            {
-                host.ReactInstanceManager.OnLeavingBackground();
-            }
-        }
-
-        /// <summary>
-        /// Applies the activation arguments.
-        /// </summary>
-        /// <param name="host">The React Native host.</param>
-        /// <param name="arguments">The arguments.</param>
-        public static void ApplyArguments(this ReactNativeAppHost host, string arguments)
-        {
-            if (!string.IsNullOrEmpty(arguments) && host.HasInstance)
-            {
-                var args = arguments.Split(',');
-
-                var index = Array.IndexOf(args, "remoteDebugging");
-                if (index < 0)
-                {
-                    return;
-                }
-
-                if (args.Length <= index + 1)
-                {
-                    throw new ArgumentException("Expected value for remoteDebugging argument.", nameof(arguments));
-                }
-
-                bool isRemoteDebuggingEnabled;
-                if (bool.TryParse(args[index + 1], out isRemoteDebuggingEnabled))
-                {
-                    if (host.HasInstance)
-                    {
-                        host.ReactInstanceManager.DevSupportManager.IsRemoteDebuggingEnabled = isRemoteDebuggingEnabled;
-                    }
-                }
-            }
-        }
-    }
-}
